refactor(discord): split message process runtime
